Add some color and switch to f64 as base number type
Lock to commit

Remove labelable trait

Upgrade from f32 to f64 to work with chrono library

Add comment and run cargo format

Add ability to a legend

Improve examples

Refactor color api to have a default without color

Switch to latest drawille

Change api for adding labels and colours

Update Cargo file to update authors and name

+++ src/main.rs
@@ -1,6 +1,6 @@
+use pointplots::{Chart, PixelColor, Plot, Shape};
 use std::process::exit;
 use structopt::StructOpt;
-use textplots::{Chart, Plot, Shape};
 
 #[derive(StructOpt)]
 struct Opt {
@@ -9,10 +9,10 @@ struct Opt {
     formula: String,
     /// X-axis start value.
     #[structopt(long, default_value = "-10.0")]
-    xmin: f32,
+    xmin: f64,
     /// X-axis end value.
     #[structopt(long, default_value = "10.0")]
-    xmax: f32,
+    xmax: f64,
     /// Canvas width in points.
     #[structopt(short, long, default_value = "180")]
     width: u32,
@@ -40,7 +40,7 @@ fn main() {
     };
 
     println!("y = {}", opt.formula);
-    Chart::<'_, f32, f32>::new(opt.width, opt.height, opt.xmin, opt.xmax)
-        .lineplot(&Shape::Continuous(Box::new(|x| func(x.into()) as f32)))
+    Chart::<'_, f64, f64>::new(opt.width, opt.height, opt.xmin, opt.xmax)
+        .lineplot(&Shape::Continuous(Box::new(|x| func(x.into()) as f64)))
         .display();
 }
